Move glyph/marker methods to a mixin
